last touches

+++ src/dev/masters/services/IService.java
@@ -16,7 +16,7 @@ public interface IService<T> {
     
     public void ajouter(T t);
     public List<T> afficher()throws SQLException;
-    public void supprimer(Long id) throws SQLException;
-    public void modifier(int id,T t) throws SQLException;
+    public void supprimer(long id) throws SQLException;
+    public void modifier(long id,T t) throws SQLException;
 
 }
